Move cheat tracking outside of control.lua
